# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments/{assignmentId}/readiness:
    get:
      operationId: getAssignmentReadiness
      tags: [analytics]
      summary: Analytics-driven readiness + surgical focus pointers for a graded assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentReadiness"
...
components:
  schemas:
...
    FocusPointer:
      type: object
      required: [topicTitle, severity, why, action]
      properties:
        topicId: { type: ["integer", "null"] }
        topicTitle: { type: string }
        severity:
          type: string
          enum: [critical, important, minor]
        why: { type: string }
        action: { type: string }
...
    AssignmentReadiness:
      type: object
      required: [assignmentId, assignmentTitle, kind, readiness, label, summary, pointers, practiceCount, recommendPractice]
      properties:
        assignmentId: { type: integer }
        assignmentTitle: { type: string }
        kind:
          type: string
          enum: [homework, test, midterm, final]
        readiness:
          type: number
          description: 0-100 readiness score for the graded assignment
        label:
          type: string
          enum: [not_ready, getting_there, ready, mastered]
        summary: { type: string }
        pointers:
          type: array
          items:
            $ref: "#/components/schemas/FocusPointer"
        practiceCount: { type: integer }
        recommendPractice: { type: boolean }
...