# lib/api-zod/src/generated/types/assignmentReadiness.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentReadinessKind } from './assignmentReadinessKind';
import type { AssignmentReadinessLabel } from './assignmentReadinessLabel';
import type { FocusPointer } from './focusPointer';

export interface AssignmentReadiness {
  assignmentId: number;
  assignmentTitle: string;
  kind: AssignmentReadinessKind;
  /** 0-100 readiness score for the graded assignment */
  readiness: number;
  label: AssignmentReadinessLabel;
  summary: string;
  pointers: FocusPointer[];
  practiceCount: number;
  recommendPractice: boolean;
}